hub/requirements.txt
paho-mqtt==1.4.0
influxdb==5.2.0